+++ src/pl/isangeles/senlin/core/item/Weapon.java
@@ -111,7 +111,7 @@ public class Weapon extends Equippable
 	 * @throws IOException
 	 * @throws FontFormatException
 	 */
-	public Weapon(String id, int serial, WeaponType type, ItemMaterial material, int value, int minDmg, int maxDmg, Modifiers bonuses, List<String> equipEffects, 
+	public Weapon(String id, long serial, WeaponType type, ItemMaterial material, int value, int minDmg, int maxDmg, Modifiers bonuses, List<String> equipEffects, 
 				  List<String> hitEffects, int reqLevel, String picName, String spriteName, GameContainer gc) 
 			throws SlickException, IOException, FontFormatException 
 	{
@@ -168,7 +168,7 @@ public class Weapon extends Equippable
 	 * @throws IOException
 	 * @throws FontFormatException
 	 */
-	public Weapon(String id, int serial, WeaponType type, ItemMaterial material, int value, int minDmg, int maxDmg, Modifiers bonuses, List<String> equipEffects, 
+	public Weapon(String id, long serial, WeaponType type, ItemMaterial material, int value, int minDmg, int maxDmg, Modifiers bonuses, List<String> equipEffects, 
 				  List<String> hitEffects, int reqLevel, String picName, GameContainer gc) throws SlickException, IOException, FontFormatException 
 	{
 		this(id, serial, type, material, value, minDmg, maxDmg, bonuses, equipEffects, hitEffects, reqLevel, picName, type.getDefaultMaleSpritesheet(), gc);
